Editor as a separate component
